+++ relm4-components/src/open_dialog.rs
@@ -10,9 +10,9 @@ use crate::ParentWindow;
 /// Configuration for the open dialog component
 pub struct OpenDialogSettings {
     /// Label for cancel button
-    pub cancel_label: String,
+    pub cancel_label: &'static str,
     /// Label for accept button
-    pub accept_label: String,
+    pub accept_label: &'static str,
     /// Allow or disallow creating folders
     pub create_folders: bool,
     /// Modal dialogs freeze other windows as long they are visible
@@ -112,8 +112,8 @@ where
             set_visible: watch!(model.is_active),
             add_filter: iterate!(&model.settings.filters),
             set_create_folders: model.settings.create_folders,
-            set_cancel_label: Some(&model.settings.cancel_label),
-            set_accept_label: Some(&model.settings.accept_label),
+            set_cancel_label: Some(model.settings.cancel_label),
+            set_accept_label: Some(model.settings.accept_label),
             set_modal: model.settings.is_modal,
             set_transient_for: parent_widgets.parent_window().as_ref(),
             connect_response => move |dialog, res_ty| {