+++ relm4-examples/examples/open_button.rs
@@ -46,8 +46,8 @@ impl OpenButtonParent for AppModel {
 
     fn dialog_config(&self) -> OpenDialogSettings {
         OpenDialogSettings {
-            accept_label: "Open".to_string(),
-            cancel_label: "Cancel".to_string(),
+            accept_label: "Open",
+            cancel_label: "Cancel",
             create_folders: true,
             is_modal: true,
             filters: Vec::new(),