+++ CHANGES.md
@@ -12,6 +12,7 @@
 
 + book: Many book improvements by [tronta](https://github.com/tronta)
 + relm4-macros: Make fields of public widgets public
++ relm4-components: Use &'static str instead of String for configurations
 + relm4-examples: Many improvements
 
 ### Fixed